requirements.txt
instaloader
mysql-connector-python
pandas==1.3.3
scipy==1.7.1
sqlalchemy
apache-airflow